Skip Active Directory tests (#6368)
