+++ google/resource_active_directory_domain_trust_test.go
@@ -10,6 +10,9 @@ import (
 )
 
 func TestAccActiveDirectoryDomainTrust_activeDirectoryDomainTrustBasicExample(t *testing.T) {
+	// skip the test until Active Directory setup issue got resolved
+	t.Skip()
+
 	// This test continues to fail due to AD setup required
 	// Skipping in VCR to allow for fully successful test runs
 	skipIfVcr(t)
